Abstract DummySpillerFactory out of Hash Aggregation Tests
